# update _winding for crs to reduce allocations