@juliohm 2024-07-02
Thank you for catching this. Perhaps we should simply merge the two branches back again? There is no need to introduce a `_winding` that dispatches on the coordinate type if all coordinates can be handled with the same technique. I believe that the call to `flat` on `Cartesian` will be free of allocations too.