@disberd 2024-07-02
@juliohm yeah if indeed calling flat by default in a single method would not compromise behavior and not create allocations/performance penalties I agree it would be better to have again the single method.

Don't know if it's worth adding a test with a dummy winding method that does not call flat to make sure that even in the future the default inclusion of `flat` does not cause more allocations than the one without flat for Cartesian.