@juliohm 2024-07-02
Will refactor the master branch to consider this change, and also add a dummy function in the tests as suggested.